Dinamizar comboboxes de VistaOP

+++ Control de calidad/src/Modelo/Color.java
@@ -11,9 +11,9 @@ package Modelo;
 public class Color {
     
     private String descripcion;
-    private long codigo;
+    private int codigo;
 
-    public Color(String descripcion, long codigo) {
+    public Color(String descripcion, int codigo) {
         this.descripcion = descripcion;
         this.codigo = codigo;
     }
@@ -26,11 +26,11 @@ public class Color {
         this.descripcion = descripcion;
     }
 
-    public long getCodigo() {
+    public int getCodigo() {
         return codigo;
     }
 
-    public void setCodigo(long codigo) {
+    public void setCodigo(int codigo) {
         this.codigo = codigo;
     }
